ENH:fe: specialize SolverMatrix to allow for more dynamic type hinting

+++ cheartpy-fe/src/cheartpy/fe/trait/_solver_group.py
@@ -64,7 +64,7 @@ class ISolverGroup(abc.ABC):
     def set_iteration(
         self,
         task: IterationSetting,
-        val: int | str,
+        val: float | str,
     ) -> None: ...
     @abc.abstractmethod
     def catch_solver_errors(
